moved logging instrumentation
