+++ agentops/logging/__init__.py
@@ -1,3 +1,4 @@
 from .config import configure_logging, logger
+from .instrument_logging import setup_print_logger, upload_logfile
 
-__all__ = ["logger", "configure_logging"]
+__all__ = ["logger", "configure_logging", "setup_print_logger", "upload_logfile"]
